requirements.txt
torch>=2.1.0
torchvision>=0.16.0
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
tqdm>=4.65.0
streamlit>=1.28.0